Move omdb api and utils to lib folder
